New commandline args interface
The coverfile and stegofile now have to be specified with the -cf and
-sf options respectively.
This simplifies arguments handling code, but most of all will allow to
ommit these arguments and use stdin/stdout in the future.

Also all bugs in the old handling code shouldn't be present in the
refactored version.