Add message compression support
Compression of the message at input (and decompression of the message at
output) is now possible.

The compression relies on `zlib`, however the library is optional and
the compression features aren't included if the library isn't found on
the system.

For now the featere is turned off during compilation as it is not very
stable.