Fix embedding with capacity less then 1 byte
Previously if the file for particular method didn't have at least 1 byte
capacity, the capacity would be rounded down to 0 so nothing would get
embedded.

This is now fixed by correctly rounding the capacity.

NOTE: The data in this case will be padded to 8 bits which might be
important when extracting.